refactor(plugins): :truck: rename auto link headings plugin
